results/rebuttal/run-20231111-034248_LLMatic-zero_shot_donnemartin-system-design-oop-finance_0_20-runs_log/finance/15/LLMatic/test_bank_accounts.py
import bank_accounts


def test_link_account():
	ba = bank_accounts.BankAccount()
	ba.link_account('123', {'name': 'Test Account', 'balance': 1000})
	assert '123' in ba.accounts


def test_import_transactions():
	ba = bank_accounts.BankAccount()
	ba.link_account('123', {'name': 'Test Account', 'balance': 1000})
	assert ba.import_transactions('123') == 'Transactions imported'


def test_get_balance():
	ba = bank_accounts.BankAccount()
	ba.link_account('123', {'name': 'Test Account', 'balance': 1000})
	assert ba.get_balance('123') == 'Account balance: 1000'


def test_get_transactions():
	ba = bank_accounts.BankAccount()
	ba.link_account('123', {'name': 'Test Account', 'balance': 1000})
	assert ba.get_transactions('123') == 'Transactions: []'

